backend/requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
email-validator
httpx
openai
python-dotenv
supabase
